Made compact and standard table view into separate widgets
